move menu.h & menu.c to API Directory, move toast notification to seperate files.
