+++ TFT/src/User/API/icon_list.inc
@@ -135,7 +135,6 @@ X_ICON (LEVELING_ON)
 X_ICON (NOZZLE_DOWN)
 X_ICON (NOZZLE_UP)
 X_ICON (LEVEL_EDGE_DISTANCE)
-X_ICON (XY_UNLOCK)
 X_ICON (TUNING)
 X_ICON (PID)
 X_ICON (EXTRUDE_100)
